generilize upgrade for any two versions
