+++ roles/openshift_excluder/molecule/playbooks/upgrade_excluders.yml
@@ -1,53 +1,69 @@
 ---
 # enable 3.4 repo only
-- hosts: openshift_excluder_upgrade_34_35
+- hosts: openshift_excluder_upgrade
   tasks:
+  - debug: var=install_version
+  - debug: var=upgrade_version
   # https://github.com/ansible/ansible-modules-extras/issues/2384
   - name: "Disable 3.3 repository"
     ini_file:
       dest: /etc/yum.repos.d/aos-qa.repo
       section: rhel-7-rhaos-3.3-rpms
       option: enabled
-      value: 0
+      value: "{{ '1' if install_version | string == '3.3' else '0' }}"
   - name: "Enable 3.4 repository"
     ini_file:
       dest: /etc/yum.repos.d/aos-qa.repo
       section: rhel-7-rhaos-3.4-rpms
       option: enabled
-      value: 1
+      value: "{{ '1' if install_version | string == '3.4' else '0' }}"
   - name: "Disable 3.5 repository"
     ini_file:
       dest: /etc/yum.repos.d/aos-qa.repo
       section: rhel-7-rhaos-3.5-rpms
       option: enabled
-      value: 0
+      value: "{{ '1' if install_version | string == '3.5' else '0' }}"
 
 # # install 3.4 excluders (repeat the same process as is carried during installation, i.e. disable -> enable)
-- hosts: openshift_excluder_upgrade_34_35
+- hosts: openshift_excluder_upgrade
   roles:
     - role: openshift_excluder
       r_openshift_excluder_action: disable
+      r_openshift_excluder_enable_docker_excluder: "{{ install_enable_docker_excluder }}"
+      r_openshift_excluder_enable_openshift_excluder: "{{ install_enable_openshift_excluder }}"
     - role: openshift_excluder
       r_openshift_excluder_action: enable
+      r_openshift_excluder_enable_docker_excluder: "{{ install_enable_docker_excluder }}"
+      r_openshift_excluder_enable_openshift_excluder: "{{ install_enable_openshift_excluder }}"
 
 # # enable 3.5 repo only
-- hosts: openshift_excluder_upgrade_34_35
+- hosts: openshift_excluder_upgrade
   tasks:
+  - name: "Enable 3.4 repository"
+    ini_file:
+      dest: /etc/yum.repos.d/aos-qa.repo
+      section: rhel-7-rhaos-3.4-rpms
+      option: enabled
+      value: "{{ '1' if upgrade_version | string == '3.4' else '0' }}"
   - name: "Enable 3.5 repository"
     ini_file:
       dest: /etc/yum.repos.d/aos-qa.repo
       section: rhel-7-rhaos-3.5-rpms
       option: enabled
-      value: 1
+      value: "{{ '1' if upgrade_version | string == '3.5' else '0' }}"
 
 # install 3.5 excluders (repeat the same process as is carried during installation, i.e. disable -> enable)
-- hosts: openshift_excluder_upgrade_34_35
+- hosts: openshift_excluder_upgrade
   roles:
     - role: openshift_excluder
       r_openshift_excluder_action: disable
       r_openshift_excluder_package_state: latest
       r_openshift_excluder_docker_package_state: latest
+      r_openshift_excluder_enable_docker_excluder: "{{ upgrade_enable_docker_excluder }}"
+      r_openshift_excluder_enable_openshift_excluder: "{{ upgrade_enable_openshift_excluder }}"
     - role: openshift_excluder
       r_openshift_excluder_action: enable
       r_openshift_excluder_package_state: latest
       r_openshift_excluder_docker_package_state: latest
+      r_openshift_excluder_enable_docker_excluder: "{{ upgrade_enable_docker_excluder }}"
+      r_openshift_excluder_enable_openshift_excluder: "{{ upgrade_enable_openshift_excluder }}"
